imelda translator
I have a bachelor's degree in English education with a GPA of 3.6. During my time as a student, I was always active in participating in seminars and campus practical activities in order to further upgrade my skills and skills. Not only have I been studying for the past 4 years, I have been an English tutor at one of the English village institutions in Pare, Kediri.